backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.7
pydantic==2.5.0
aiofiles==23.2.1
python-dotenv==1.0.0
google-generativeai==0.3.2
reportlab==4.4.4
sqlalchemy==2.0.44
